# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: zent-logbook-ms
  description: Logbook microservice
  version: 1.0.0
servers:
- url: /rest/zent-logbook-api/v1.0/
paths:
  /generate_report_form:
    get:
      tags:
      - Reports
      summary: Genera un archivo en excel con los datos de feria
      description: Genera el archivo excel
      operationId: get_excel_form
      parameters:
        - name: externalTransactionId
          in: header
          required: true
          schema:
            type: string
            example: "fcea920f7412b5da7be0cf42b8c93759"
        - name: channel
          in: header
          required: true
          schema:
            type: string
            example: "ZENTINEL"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.user_controller.UserView
  /send-qr/{id_form}:
    get:
      tags:
      - Expo
      summary: Envia el correo con el qr para los vip
      description: Envia qr
      operationId: send_email_qr
      parameters:
        - name: id_form
          in: path
          description: id del formulario
          required: true
          schema:
            type: integer
            format: int64
        - name: externalTransactionId
          in: header
          required: true
          schema:
            type: string
            example: "fcea920f7412b5da7be0cf42b8c93759"
        - name: channel
          in: header
          required: true
          schema:
            type: string
            example: "ZENTINEL"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.user_controller.UserView
  /validate-qr-invitation:
    get:
      tags:
      - Expo
      summary: Obtiene todos los registros del evento
      description: Devuelve todos los registros de la base
      operationId: validate_qr
      parameters:
        - name: qr-token
          in: query
          description: token del qr
          required: true
          schema:
            type: string
            format: uuid
        - name: externalTransactionId
          in: header
          required: true
          schema:
            type: string
            example: "fcea920f7412b5da7be0cf42b8c93759"
        - name: channel
          in: header
          required: true
          schema:
            type: string
            example: "ZENTINEL"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.user_controller.UserView    
  /get/all-users:
    get:
      tags:
      - Users
      summary: Obtiene todos los usuarios
      description: Devuelve todos los usuarios de la base
      operationId: get_all_users
      parameters:
        - name: externalTransactionId
          in: header
          required: true
          schema:
            type: string
            example: "fcea920f7412b5da7be0cf42b8c93759"
        - name: channel
          in: header
          required: true
          schema:
            type: string
            example: "ZENTINEL"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.user_controller.UserView
...
  /post/new-user:
    post:
      tags:
      - Users
      summary: Guarda el usuario en la base de datos.
      description: Guardado de usuario
      operationId: new_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPostNewUser"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
      x-openapi-router-controller: swagger_server.controllers.user_controller.UserView
...
components:
  schemas:
...
    RequestPostNewUser:
      properties:
        new_user:
          $ref: "#/components/schemas/NewUserData"
        channel:
          type: string
          example: ZENTINEL
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
...
    NewUserData:
      type: object
      properties:
        user:
          type: string
          example: "dmales"
        email:
          type: string
          example: "daniel@hotmail.com"
        password:
          type: string
          example: "123456"
        attributes:
          type: object
        role_id:
          type: string
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
      example:
        user: "dmales"
        email: "daniel@hotmail.com"
        password: "123456"
        attributes: {}
        role_id: "d290f1ee-6c54-4b01-90e6-d701748f0851"
    GenericResponse:
      properties:
        data:
          type: object
          nullable: true
          example: Estructura generica.
        errorCode:
          type: integer
          example: 0
        message:
          type: string
          example: Mensaje generico.
        internalTransactionId:
          type: string
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId:
          type: string
          example: fcea920f7412b5da7be0cf42b8c93759
      example:
        errorCode: 0
        internalTransactionId: d290f1ee-6c54-4b01-90e6-d701748f0851
        externalTransactionId: fcea920f7412b5da7be0cf42b8c93759
        message: Mensaje generico.
        data: {}
    ResponseError:
      type: object
      properties:
        errorCode:
          type: integer
        internalTransactionId:
          type: string
        externalTransactionId:
          type: string
        message:
          type: string
...